requirements.txt
Flask==2.0.1
flask-cors==3.0.10
Werkzeug==2.0.3
python-dotenv==0.19.0
newsapi-python==0.2.6
requests==2.26.0 